+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/shell/Shell.scala
@@ -41,11 +41,11 @@ import com.dimajix.flowman.JAVA_VERSION
 import com.dimajix.flowman.SCALA_VERSION
 import com.dimajix.flowman.SPARK_BUILD_VERSION
 import com.dimajix.flowman.SPARK_VERSION
+import com.dimajix.flowman.common.ConsoleColors.yellow
 import com.dimajix.flowman.common.Logging
 import com.dimajix.flowman.common.ParserUtils.splitSettings
 import com.dimajix.flowman.common.ToolConfig
 import com.dimajix.flowman.tools.StatefulTool
-import com.dimajix.flowman.util.ConsoleColors.yellow
 
 
 object Shell {